раскидал бузыку и звуки, правка кода

+++ Assets/Scripts/Sound/SoundAction.cs
@@ -7,8 +7,8 @@ public class SoundAction : MonoBehaviour
 
     private void Awake()
     {
-        _audioSourceMusic.volume = PlayerPrefs.GetInt(Enums.AudioName.Music.ToString());
-        int currentValue = PlayerPrefs.GetInt(Enums.AudioName.Effect.ToString());
+        _audioSourceMusic.volume = PlayerPrefs.GetFloat(Enums.AudioName.Music.ToString());
+        float currentValue = PlayerPrefs.GetFloat(Enums.AudioName.Effect.ToString());
 
         foreach (AudioSource audioSource in _audioSources)
         {
